Cleanup warnings, add logging

+++ src/async_web_server_cpp/websocket_message.cpp
@@ -194,7 +194,7 @@ std::optional<bool> WebsocketFrameParser::consume(WebsocketFrame& frame, char in
     //unmask the frame
     if (frame.header.mask)
     {
-      for (int i = 0; i < frame.length; ++i)
+      for (size_t i = 0; i < frame.length; ++i)
       {
         frame.content[i] = frame.content[i] ^ frame.mask[i % 4];
       }
